Add jekyll-feed support
This allows for GitHub Pages to automatically generate an atom feed for
the site using https://github.com/jekyll/jekyll-feed. This also adds the
feed_meta to the head so it can be autodiscovered, as well as sets the
name in the config as the feed_meta uses the name for the feed title
instead of the Jekyll title.