Add support for YouTube embeds
If a post has a youtube_id, then the video gets embedded. It uses the
no-cookie embed so a cooke is only added if the video is played, and it
assumes the max-width of the site and that the video is 16:9. No
suggested videos are displayed.

The added styles are from
https://css-tricks.com/NetMag/FluidWidthVideo/Article-FluidWidthVideo.php
to make the embedded video responsive (with some adjustments). It only
supports 16:9, which is totally fine.